requirements.txt
langchain-google-genai
python-gitlab
pyyaml
gitpython
